requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
langchain==0.2.10
langchain-community==0.2.9
langchain-openai==0.1.17
langchain-text-splitters==0.2.1
langchain-huggingface==0.0.3
chromadb==0.5.3
pypdf==4.2.0
sentence-transformers==2.7.0
